frontend: Rewrite controller infrastructure
This pulls out the AppContext from the direct call tree into the
request where it makes more sense. Also it's not really idiomatic
to use it directly, so some parts got rewritten in this (and the
previous) commit to use static dependency declarations within the
constructor so it all works out better.