client: Remove improperly implemented strandOn boolean
Technically it shouldn't even be available, since the strand would
have random lights switch on close to its start because of noise
or something, therefore we power it off occasionally. Having once
flushed to it won't help proactively, so if you want to display a
static color, refresh it periodically instead.